test: move conftest in order to prevent unit tests from trying to use it
