+++ tests/functional/clients/test_enhance_client.py
@@ -5,7 +5,7 @@ import pytest
 from aiohttp import ClientResponseError
 from diffbot_kg.clients.enhance import DiffbotEnhanceClient
 
-from tests.functional.clients.conftest import (
+from tests.functional.conftest import (
     ORG2_ENTITY_ID,
     ORG2_NAME,
     ORG2_URL,
