+++ tests/functional/clients/test_search_client.py
@@ -1,7 +1,7 @@
 import pytest
 from diffbot_kg.clients.search import DiffbotSearchClient
 
-from tests.functional.clients.conftest import ORG_ENTITY_ID, ORG_NAME, Secret
+from tests.functional.conftest import ORG_ENTITY_ID, ORG_NAME, Secret
 
 
 @pytest.fixture(scope="session")
